Changed the algorithm
At first the algorith was the default one for evaluating a Polish
notation strings. However, the algorith was working with only two
operands per operator. This leads to problmes when for instance one want
to calculate + 2 3 4 5 6

This commit enhances the original alghoritm by adding a little bit more
simple logic which is used and is working with more than two operands
per operator